% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bc_wave.R
\name{bc.cwt}
\alias{bc.cwt}
\title{Title}
\usage{
bc.cwt(
  data,
  subset = 1:256,
  method = c("additive", "multiplicative"),
  variable,
  theta = 0.1,
  QM = TRUE,
  PRand = FALSE,
  number_sim = 1,
  wavelet = "morlet",
  dt = 1,
  dj = 1/8,
  upperPeriod,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
