% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fun_cwt_J}
\alias{fun_cwt_J}
\title{Function: Total number of decomposition levels}
\usage{
fun_cwt_J(n, dt, dj)
}
\arguments{
\item{dj}{}
}
\description{
Function: Total number of decomposition levels
}
