% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fun_ifft}
\alias{fun_ifft}
\title{Inverse Fourier transform}
\usage{
fun_ifft(x, do.plot = FALSE)
}
\arguments{
\item{x}{input time series.}

\item{do.plot}{Logical value of plot.}
}
\value{
reconstruction time series
}
\description{
Inverse Fourier transform
}
\examples{
x <- rnorm(100)
x.new <- fun_ifft(x, do.plot=TRUE)
}
\references{
fun_stoch_sim in PRSim, Brunner and Furrer, 2020.
}
