% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{NWP.rain}
\alias{NWP.rain}
\title{Australia NWP rainfall forecasts at lead 1h over Sydney region}
\usage{
data(NWP.rain.h)
}
\description{
A dataset containing 160 stations including observation and raw forecasts.
}
\keyword{datasets}
