% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_verification.R
\name{RankHist}
\alias{RankHist}
\title{Verification Rank and Histogram}
\usage{
RankHist(forecasts, observations, do.plot = FALSE)
}
\arguments{
\item{forecasts}{A matrix of ensemble forecasts, in which the rows corresponds to locations and times and the columns correspond to the individual ensemble members.}

\item{observations}{A vector of observations corresponding to the locations and times of the forecasts.}

\item{do.plot}{Logical value of plot.}
}
\value{
A vector giving the rank of verifying observations relative to the corresponding ensemble forecasts. The verification rank historgram is plotted.
}
\description{
Verification Rank and Histogram
}
\references{
ensembleBMA::verifRankHist
}
