% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sample}
\alias{sample}
\title{Sample data: Rainfall forecasts data}
\usage{
data(sample)
}
\description{
A dataset containing 2 stations including observation and raw forecasts.
}
\keyword{datasets}
