% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bc_cwt.R
\name{bc_cwt}
\alias{bc_cwt}
\title{bc_cwt}
\usage{
bc_cwt(
  data,
  subset,
  variable,
  theta = 0.1,
  QM = c("MBC", "MRS", "QDM"),
  number_sim = 5,
  wavelet = "morlet",
  dt = 1,
  dj = 1,
  J = 9,
  PR.cal = FALSE,
  do.plot = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
bc_cwt
}
