% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fun_cwt_J}
\alias{fun_cwt_J}
\title{Function: Total number of decomposition levels}
\usage{
fun_cwt_J(n, dt, dj)
}
\arguments{
\item{n}{sample size.}

\item{dt}{sampling resolution in the time domain.}

\item{dj}{sampling resolution in the frequency domain.}
}
\value{
the total number of decomposition levels.
}
\description{
Function: Total number of decomposition levels
}
