% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fun_icwt}
\alias{fun_icwt}
\title{Function: Inverse of continuous wavelet transform}
\usage{
fun_icwt(x.wave, dt = 1, dj = 1/8, lowerPeriod = 2 * dt, scales = NULL)
}
\arguments{
\item{x}{input complex matrix}
}
\value{
reconstruction time series
}
\description{
Function: Inverse of continuous wavelet transform
}
